Add sanke Hopping ability
