+++ Assets/Animation/Snake/SnakeHop.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d38f40223e6b741188eae3b4c7e548e3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
