+++ Assets/Animation/Snake/SnakeShoot.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a82669da1ef44e43b680c368071ceb8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
