--- Assets/Scripts/SnakeSpitter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a9cef9e9334c14d7b93e000a6a79262b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
